feat (DPLAN-11470) adjust test to have common code in on class
